Move grammar model classes to their own package to distinguish them from other application model classes.
