+++ consent-ws/src/main/java/org/genomebridge/consent/http/models/Consent.java
@@ -1,6 +1,7 @@
 package org.genomebridge.consent.http.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.genomebridge.consent.http.models.grammar.UseRestriction;
 
 /**
  * Consent Representation object.
