+++ consent-ws/src/main/java/org/genomebridge/consent/http/models/grammar/And.java
@@ -1,23 +1,20 @@
-package org.genomebridge.consent.http.models;
+package org.genomebridge.consent.http.models.grammar;
 
 import com.google.common.base.Objects;
 
 import java.util.Arrays;
 
-public class Or extends UseRestriction {
+public class And extends UseRestriction {
 
-    private String type = "or";
+    private String type = "and";
 
     private UseRestriction[] operands;
 
-    public Or() {
+    public And() {
     }
 
-    public Or(UseRestriction... operands) {
+    public And(UseRestriction... operands) {
         this.operands = operands;
-        if (operands.length < 2) {
-            throw new IllegalArgumentException("Disjunction must have at least two operands");
-        }
     }
 
     public String getType() {
@@ -39,8 +36,8 @@ public class Or extends UseRestriction {
 
     @Override
     public boolean equals(Object o) {
-        return o instanceof Or &&
-                Arrays.deepEquals(this.operands, ((Or) o).operands);
+        return o instanceof And &&
+                Arrays.deepEquals(this.operands, ((And) o).operands);
     }
 
     public boolean visitAndContinue(UseRestrictionVisitor visitor) {
