+++ consent-ws/src/main/java/org/genomebridge/consent/http/models/grammar/Named.java
@@ -1,4 +1,4 @@
-package org.genomebridge.consent.http.models;
+package org.genomebridge.consent.http.models.grammar;
 
 import com.google.common.base.Objects;
