+++ consent-ws/src/main/java/org/genomebridge/consent/http/models/grammar/UseRestrictionVisitor.java
@@ -1,4 +1,4 @@
-package org.genomebridge.consent.http.models;
+package org.genomebridge.consent.http.models.grammar;
 
 public interface UseRestrictionVisitor {
