+++ consent-ws/src/test/java/org/genomebridge/consent/http/ConsentsServiceTest.java
@@ -8,7 +8,7 @@ import io.dropwizard.testing.junit.DropwizardAppRule;
 import org.apache.commons.lang.StringUtils;
 import org.genomebridge.consent.http.models.Consent;
 import org.genomebridge.consent.http.models.ConsentAssociation;
-import org.genomebridge.consent.http.models.Everything;
+import org.genomebridge.consent.http.models.grammar.Everything;
 import org.junit.ClassRule;
 import org.junit.Test;
